Component Separation Done according to diagram
