+++ src/components/Form/Form.module.css
@@ -1,14 +1,3 @@
-.main {
-  --padding-top: 1rem;
-  padding-top: var(--padding-top);
-  width: 100%;
-  height: calc(100vh - var(--padding-top));
-  display: flex;
-  flex-flow: column nowrap;
-  align-items: center;
-}
-
-/* Form Styles */
 .form {
   display: flex;
   flex-flow: column nowrap;
@@ -27,6 +16,10 @@
   display: flex;
   flex-flow: row wrap;
   justify-content: space-between;
+}
+
+.input-group article label {
+  color: var(--main-text-color);
 }
 
 .input-group article input {
