+++ src/components/UI/Button/Button.module.css
@@ -1,0 +1,17 @@
+.button {
+  padding: 0.6rem;
+  cursor: pointer;
+  border-radius: 0.3rem;
+}
+
+.button.primary {
+  background-color: #8000b3;
+  border: none;
+  color: var(--main-text-color);
+}
+
+.button.secondary {
+  background-color: #414A4C;
+  border: none;
+  color: var(--main-text-color);
+}
