+++ src/index.css
@@ -1,6 +1,11 @@
+:root {
+  --main-text-color: #f2f2f2;
+}
+
 * {
   margin: 0;
   padding: 0;
+  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
 }
 
 body {
